requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
chromedriver-autoinstaller==0.6.4
click==8.1.8
cssselect==1.2.0
dataclasses-json==0.6.7
Deprecated==1.2.15
deptry==0.21.2
dirtyjson==1.0.8
elastic-transport==8.15.1
elasticsearch==8.17.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jieba3k==0.35.1
joblib==1.4.2
llama-index-core==0.12.8
llama-index-embeddings-llamafile==0.3.0
llama-index-embeddings-ollama==0.5.0
llama-index-llms-llamafile==0.3.0
llama-index-llms-ollama==0.5.0
llama-index-readers-file==0.4.1
llama-index-readers-web==0.3.3
llama-index-vector-stores-elasticsearch==0.4.0
lxml==5.3.0
marshmallow==3.23.2
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.1
ollama==0.4.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
playwright==1.49.1
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
pyee==12.0.0
pypdf==5.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requirements-parser==0.11.0
selenium==4.27.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spider-client==0.0.27
SQLAlchemy==2.0.36
striprtf==0.0.26
tenacity==9.0.0
tiktoken==0.8.0
tinysegmenter==0.3
tldextract==5.1.3
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
types-setuptools==75.6.0.20241223
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websocket-client==1.8.0
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
